Completed tests for construction of L, F and T for Query-Target
